requirements.txt
```txt
python-dotenv
unidecode
openai
python-telegram-bot
gspread
google-oauth2-service-account
schedule
langdetect
Levenshtein
```